requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0

# PDF Parsing
PyMuPDF>=1.23.0
pypdf>=4.0.0

# OCR (optional fallback)
pytesseract>=0.3.10
Pillow>=10.0.0

# Data handling
pandas>=2.1.0
openpyxl>=3.1.2
numpy>=1.26.0

# Embeddings / Similarity
scikit-learn>=1.4.0
faiss-cpu>=1.7.4

# FastAPI file upload
python-multipart>=0.0.9
